refactor: msw api 랜덤 값 만드는 로직 재사용, 리액트 쿼리 staleTime, gcTime 조정
